Add blinking text support (SGR 5)
GhosttyStyle already tracked the blink flag; wire it through the
snapshot (attrs array widened from byte[] to int[] to fit the new
ATTR_BLINK = 32 bit) and animate it in TerminalView with a 600 ms
postDelayed ticker — the same pattern used for cursor blink. The
ticker starts on demand when any viewport cell carries ATTR_BLINK
and stops automatically when none do.

+++ app/src/main/java/sh/easycli/proot/term/ScreenSnapshot.java
@@ -21,7 +21,7 @@ public final class ScreenSnapshot {
     public int[] codepoints = new int[0];
     public int[] fg = new int[0];
     public int[] bg = new int[0];
-    public byte[] attrs = new byte[0];
+    public int[] attrs = new int[0];
     /** See terminal_jni.c terminalSnapshot for the layout. */
     public final int[] meta = new int[16];
 
@@ -106,7 +106,7 @@ public final class ScreenSnapshot {
         codepoints = new int[cells];
         fg = new int[cells];
         bg = new int[cells];
-        attrs = new byte[cells];
+        attrs = new int[cells];
     }
 
     /** Grows the grapheme buffer to hold {@code ints} record ints plus slot 0. */
